rename to words.ts
