+++ src/utils/words.ts
